For CHINESE v0.2 - U7! Repeatable position scaling fixed!

+++ src/rs/alexanderstojanovich/fo2ie/util/GLCoords.java
@@ -33,8 +33,8 @@ public class GLCoords {
      * @return OpenGL coordinates
      */
     public static Vector2f getOpenGLCoordinates(Vector2f scrnCoords, int scrnWidth, int scrnHeight) {
-        float xposGL = (float) (scrnCoords.x / scrnWidth - 0.5f) * 2.0f;
-        float yposGL = (float) (0.5f - scrnCoords.y / scrnHeight) * 2.0f;
+        float xposGL = (float) (scrnCoords.x / (float) scrnWidth - 0.5f) * 2.0f;
+        float yposGL = (float) (0.5f - scrnCoords.y / (float) scrnHeight) * 2.0f;
 
         return new Vector2f(xposGL, yposGL);
     }
